aienginev2: src: transaction mode of operation
Implementation of transaction mode for the ai engine driver.
In transaction mode, the driver stores all the IO operations in an
internally managed buffer until an API results in auto flush or the user
explicitly calling submit transaction. Each transaction instance is
associated with the thread id of the execution context. Flags are exposed to
the user to enable or disable auto-flush mode.

Signed-off-by: Tejus Siddagangaiah <tejus.siddagangaiah@xilinx.com>

Acked-for-series: Wendy Liang <wendy.liang@xilinx.com>