aienginev2: src: npi: remove the columns checking for protected reg
As the NPI access are through internal APIs, it is not required to
check the columns information in the request. As for aie device, the
enable protected registers NPI register doesn't have the columns
field, the API can just ignore the columns information.

Signed-off-by: Wendy Liang <wendy.liang@xilinx.com>

Acked-for-series: Tejus Siddagangaiah <tejus.siddagangaiah@xilinx.com>