driver: src: Corrected setting BurstLen for 512B in DmaSetAxi API
Current logic of setting Burst length in BD is not correct for 512B,
this patch corrects the same.
Assigns null pointer to the burstlencheck function pointer in DmaMod
structure initialisations for MemTile and AieTile as is not required
for the same.
Also moved BurstLenCheck API from xaie_dma_aie2ps.c to xaie_dma_aie2p.c
as this is required from aie2p architecture.

Signed-off-by: Jagadeesh Banisetti <jagadeesh.banisetti@amd.com>

Acked-for-series: Sandip Kothari <sandip.kothari@amd.com>